worker/requirements.txt
supabase==1.0.3
python-dotenv==1.0.1
loguru==0.7.2
requests
urllib3<2.0.0